+++ TGA_FTIR_tools/input_output/TGA.py
@@ -18,7 +18,7 @@ POLYORDER = int(SAVGOL.getfloat("POLYORDER"))
 def TGA_info(file, TGA, profile=COUPLING["profile"]):
     from ..classes import SampleInfo
 
-    profile = read_profile_json(profile)['tga']
+    profile = read_profile_json(profile)["data"]['tga']
 
     "extract TG info e.g. measurement time, initial mass... from TG file"
     # open file from TGA in given directory and make a DataFrame from it
@@ -50,123 +50,87 @@ def TGA_info(file, TGA, profile=COUPLING["profile"]):
 
 def default_info(name, tga):
     from ..classes import SampleInfo
-
     info = SampleInfo(
         name=name,
-        initial_mass=tga.loc[0, "sample_mass"] if "sample_mass" in tga.columns else None,
-        step_temp=[max(tga["reference_temp"])],
+        initial_mass=tga["sample_mass"].iloc[0] if "sample_mass" in tga.columns else None,
+        final_mass = tga["sample_mass"].iloc[-1] if "sample_mass" in tga.columns else None,
+        steps_idx = {"initial":0, "final":tga.index.size-1}
     )
     return info
 
 
-def dry_weight(
-    sample, how_dry="H2O", ref_mass="dry_mass",
-):
-
+def dry_weight(sample, how_dry="H2O"):
     "determine dry point and mass from TG data"
-    # make h2o to H2O
-    if how_dry == "h2o":
-        how_dry = how_dry.upper()
 
     # if how_dry is None, no dry point is determined
     if how_dry == None:
-        dry_point = 1
+        dry_point_idx = 0
 
-    # if how_dry is a number, the dry point is set to that temperature
-    elif type(how_dry) != str:
-        # check if value is within the temperature range
-        if (how_dry > max(sample.tga["sample_temp"])) or (
-            how_dry <= min(sample.tga["sample_temp"])
-        ):
-            how_dry = None
-            dry_point = 1
-            logger.warning(
-                "Supplied value is out of range. 'how_dry' is set to None. Be aware that 'reference_mass' is also set to 'initial_mass'."
-            )
-        else:
-            dry_point = sample.tga["time"][sample.tga["sample_temp"] >= how_dry].values[
-                0
-            ]
-            # check if how_dry value could not be found (very scarce, but possible..)
-            if dry_point == 0:
-                how_dry = None
-                dry_point = 1
-                logger.warning(
-                    "Supplied value is out of range. 'how_dry' is set to None. Be aware that 'reference_mass' is also set to 'initial_mass'."
+    if how_dry=="H2O" and sample.ega is None:
+        logger.error(f"If {how_dry=!r}, EGA data is required.")
+
+    if col:="sample_temp" not in sample.ega.columns:
+        logger.error(f"{col!r} must be present in EGA data.")
+        return
+
+    min_temp, max_temp = sample.tga["sample_temp"].min() , sample.tga["sample_temp"].max()
+    
+    match how_dry:
+        # if how_dry is a number, the dry point is set to that temperature
+        case float() | int():
+            # check if value is within the temperature range
+            if (min_temp < how_dry < max_temp):
+                dry_point_idx = sample.tga["time"][sample.tga["sample_temp"] >= how_dry].argmax()
+            else:
+                logger.error(f"Supplied value is out of range. Must be within {min_temp:.2f} to {max_temp:.2f}")
+                return 
+                # check if how_dry value could not be found (very scarce, but possible..)
+
+        # if how_dry is 'H2O' or 'sample_mass', the dry point is determined from the respective data
+        case str():
+            if how_dry == "H2O":
+                ref = sample.ega.copy()#.filter(items=["sample_temp", "H2O"])
+            elif how_dry == "sample_mass":
+                ref = sample.tga.copy()#.filter(items=["sample_temp", "sample_mass"])
+                ref["sample_mass"] = -sp.signal.savgol_filter(
+                    sample.tga["sample_mass"] / sample.info.initial_mass,
+                    WINDOW_LENGTH,
+                    POLYORDER,
+                    deriv=1,
                 )
+            else:
+                logger.info(f"'H2O' not in EGA-data. Defaulting to '{how_dry}")
+            
+            # look for signal peak between 50 and 200 °C
+            peak_signal_idx = ref[how_dry][(50 < ref["sample_temp"]) & (ref["sample_temp"] < 200)].argmax()
+            min_T = ref["sample_temp"].iloc[peak_signal_idx]
+            max_T = min_T + 50
+            range_T = (min_T < ref["sample_temp"]) & (ref["sample_temp"]< max_T)
 
-    # if how_dry is 'H2O' or 'sample_mass', the dry point is determined from the respective data
-    else:
-        if how_dry == "H2O":
-            try:
-                ref = sample.ega.filter(items=["sample_temp", "H2O"])
-            except:
-                how_dry = "sample_mass"
+            # fit line to slope of peak and find intersection with temperature signal
+            x = ref["sample_temp"][range_T]
+            y = ref[how_dry][range_T]
+            slope, intercept, _, _, _ = sp.stats.linregress(x, y)
 
-        if how_dry == "sample_mass":
-            ref = sample.tga.filter(items=["sample_temp", "sample_mass"])
-            ref["sample_mass"] = -sp.signal.savgol_filter(
-                sample.tga["sample_mass"] / sample.info.initial_mass,
-                WINDOW_LENGTH,
-                POLYORDER,
-                deriv=1,
-            )
-
-        min_T = ref["sample_temp"][
-            ref[how_dry]
-            >= max(ref[how_dry][(ref["sample_temp"] > 50) & (ref["sample_temp"] < 200)])
-        ].values[0]
-        max_T = min_T + 50
-
-        x = ref["sample_temp"][
-            (ref["sample_temp"] > min_T) & (ref["sample_temp"] < max_T)
-        ]
-        y = ref[how_dry][(ref["sample_temp"] > min_T) & (ref["sample_temp"] < max_T)]
-        slope, intercept, _, _, _ = sp.stats.linregress(x, y)
-
-        dry_point = ref["sample_temp"][ref["sample_temp"] >= -intercept / slope].index[
-            0
-        ]
+            intersection = (ref["sample_temp"] >= -intercept / slope)
+            if intersection.any():
+                dry_point_idx = ref["sample_temp"][intersection].argmax()
+            else:
+                logger.error(f"Unable to determine dry point from {how_dry}-trace.")
+                return
+        
+        case _:
+            logger.error(f"{how_dry=!r} is of the wrong type. Must be either a number between {min_temp:.2f} to {max_temp:.2f} or one of ['H2O' and 'sample_mass'].")
+            pass
 
     # getting the dry_mass at the dry_point as well as the final weight and calculating the relative
     # mass-loss and the water content from it
-    dry_temp = sample.tga["sample_temp"][dry_point]
+    dry_temp = sample.tga["sample_temp"][dry_point_idx]
     info = {}
 
-    if not how_dry:
-        info['reference_mass'] = "initial_mass"
-        info['initial_mass'] = sample.info['initial_mass']
-        times = [0] + list(
-            sample.tga.index[sample.tga["sample_temp"].isin(sample.info.step_temp)]
-        )
-        names = sample.info.mass_steps
-
-    elif (
-        (how_dry == "H2O") or (how_dry == "sample_mass") or (type(how_dry) != str)
-    ):  # turns to be true everytime
-        times = [0, dry_point] + list(
-            sample.tga.index[sample.tga["sample_temp"].isin(sample.info.step_temp)]
-        )
-        names = ["dry"] + sample.info.mass_steps
-        info['reference_mass'] = "dry_mass"
-        info['dry_mass'] = sample.tga["sample_mass"][dry_point]
-        info['dry_temp'] = dry_temp
-        info['dry_time'] = dry_point
-
-    if ref_mass != "dry_mass":
-        info['reference_mass'] = ref_mass
-    weights = sample.tga["sample_mass"][sample.tga.index.isin(times)].values
-    mass_loss = abs(np.diff(weights))
-
-    info['reference'] = sample.reference
-
-    info['final_mass'] = sample.tga["sample_mass"][len(sample.tga) - 1]
-    for name, ml in zip(names, mass_loss):
-        info["ML_" + name] = ml
-        info["rel_ML_" + name] = ml / info[ref_mass]
-
-    info['step_temp'] = sample.tga["sample_temp"][sample.tga.index.isin(times)].unique()
-    info['step_time'] = times
-    info['mass_steps'] = names
+    info["steps_idx"] = {"dry":dry_point_idx}
+    info['reference_mass_name'] = "dry_mass"
+    info['dry_mass'] = sample.tga["sample_mass"][dry_point_idx]
+    info['dry_temp'] = dry_temp
     sample.info.update(info)