+++ TGA_FTIR_tools/input_output/__init__.py
@@ -1,3 +1,3 @@
-from .general import read_data, time
+from .general import read_data, time, read_profile_json
 from .mass_step import mass_step
 from .samplelog import samplelog